Add files via upload
Perspectives of people throughout many surveys conducted in Africa has proven that digital technology has a positive impact on a country’s sustainable economic growth, but government and business interventions have not actualized their expectations completely. This is due to the gap between well skilled individuals and individuals without the knowledge of digital technologies. By providing Digital Solutions, implementation of skills learnt will come easily. Women and youth will be inspired to do more in Ghana and in Africa as a whole.